Move rotor velocity module up to floris.simulation
